+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## []
+
+- Fixed loot overflow in REI
+- Fixed loot overflow in JEI
+
+## [1.1.4]
+
+- Fixed crash when used on standalone server
+
+## [1.1.3]
+
+- Fixed crash when used on standalone server
+- Fixed crash when failed to encode loot table
+
 ## [1.1.2]
 
 - Used compression for transferring loot tables